:wrench: Add Solidity linting

+++ test/MoonInk.t.sol
@@ -3,52 +3,53 @@ pragma solidity ^0.8.13;
 
 import "forge-std/Test.sol";
 
-import "../src/MoonInk.sol";
 import {ERC721TokenReceiver} from "solmate/tokens/ERC721.sol";
 
+import "../src/MoonInk.sol";
+
 contract MoonInkTest is Test, ERC721TokenReceiver {
-    
+    //
     MoonInk private moon;
 
     // cycle
-    uint time0 = 1652716800; // Mon May 16 2022 16:00:00 GMT+0000
-    uint time1 = 1652803200; // Tue May 17 2022 16:00:00 GMT+0000
-    uint time2 = 1652889600; // Wed May 18 2022 16:00:00 GMT+0000
-    uint time3 = 1652976000; // Thu May 19 2022 16:00:00 GMT+0000
-    uint time4 = 1653062400; // Fri May 20 2022 16:00:00 GMT+0000
-    uint time5 = 1653148800; // Sat May 21 2022 16:00:00 GMT+0000
-    uint time6 = 1653235200; // Sun May 22 2022 16:00:00 GMT+0000
-    uint time7 = 1653321600; // Mon May 23 2022 16:00:00 GMT+0000
-    uint time8 = 1653408000; // Tue May 24 2022 16:00:00 GMT+0000
-    uint time9 = 1653494400; // Wed May 25 2022 16:00:00 GMT+0000
-    uint time10 = 1653580800; // Thu May 26 2022 16:00:00 GMT+0000
-    uint time11 = 1653667200; // Fri May 27 2022 16:00:00 GMT+0000
-    uint time12 = 1653753600; // Sat May 28 2022 16:00:00 GMT+0000
-    uint time13 = 1653840000; // Sun May 29 2022 16:00:00 GMT+0000
-    uint time14 = 1653926400; // Mon May 30 2022 16:00:00 GMT+0000
-    uint time15 = 1654012800; // Tue May 31 2022 16:00:00 GMT+0000
-    uint time16 = 1654099200; // Wed Jun 01 2022 16:00:00 GMT+0000
-    uint time17 = 1654185600; // Thu Jun 02 2022 16:00:00 GMT+0000
-    uint time18 = 1654272000; // Fri Jun 03 2022 16:00:00 GMT+0000
-    uint time19 = 1654358400; // Sat Jun 04 2022 16:00:00 GMT+0000
-    uint time20 = 1654444800; // Sun Jun 05 2022 16:00:00 GMT+0000
-    uint time21 = 1654531200; // Mon Jun 06 2022 16:00:00 GMT+0000
-    uint time22 = 1654617600; // Tue Jun 07 2022 16:00:00 GMT+0000
-    uint time23 = 1654704000; // Wed Jun 08 2022 16:00:00 GMT+0000
-    uint time24 = 1654790400; // Thu Jun 09 2022 16:00:00 GMT+0000
-    uint time25 = 1654876800; // Fri Jun 10 2022 16:00:00 GMT+0000
-    uint time26 = 1654963200; // Sat Jun 11 2022 16:00:00 GMT+0000
-    uint time27 = 1655049600; // Sun Jun 12 2022 16:00:00 GMT+0000
+    uint256 time0 = 1652716800; // Mon May 16 2022 16:00:00 GMT+0000
+    uint256 time1 = 1652803200; // Tue May 17 2022 16:00:00 GMT+0000
+    uint256 time2 = 1652889600; // Wed May 18 2022 16:00:00 GMT+0000
+    uint256 time3 = 1652976000; // Thu May 19 2022 16:00:00 GMT+0000
+    uint256 time4 = 1653062400; // Fri May 20 2022 16:00:00 GMT+0000
+    uint256 time5 = 1653148800; // Sat May 21 2022 16:00:00 GMT+0000
+    uint256 time6 = 1653235200; // Sun May 22 2022 16:00:00 GMT+0000
+    uint256 time7 = 1653321600; // Mon May 23 2022 16:00:00 GMT+0000
+    uint256 time8 = 1653408000; // Tue May 24 2022 16:00:00 GMT+0000
+    uint256 time9 = 1653494400; // Wed May 25 2022 16:00:00 GMT+0000
+    uint256 time10 = 1653580800; // Thu May 26 2022 16:00:00 GMT+0000
+    uint256 time11 = 1653667200; // Fri May 27 2022 16:00:00 GMT+0000
+    uint256 time12 = 1653753600; // Sat May 28 2022 16:00:00 GMT+0000
+    uint256 time13 = 1653840000; // Sun May 29 2022 16:00:00 GMT+0000
+    uint256 time14 = 1653926400; // Mon May 30 2022 16:00:00 GMT+0000
+    uint256 time15 = 1654012800; // Tue May 31 2022 16:00:00 GMT+0000
+    uint256 time16 = 1654099200; // Wed Jun 01 2022 16:00:00 GMT+0000
+    uint256 time17 = 1654185600; // Thu Jun 02 2022 16:00:00 GMT+0000
+    uint256 time18 = 1654272000; // Fri Jun 03 2022 16:00:00 GMT+0000
+    uint256 time19 = 1654358400; // Sat Jun 04 2022 16:00:00 GMT+0000
+    uint256 time20 = 1654444800; // Sun Jun 05 2022 16:00:00 GMT+0000
+    uint256 time21 = 1654531200; // Mon Jun 06 2022 16:00:00 GMT+0000
+    uint256 time22 = 1654617600; // Tue Jun 07 2022 16:00:00 GMT+0000
+    uint256 time23 = 1654704000; // Wed Jun 08 2022 16:00:00 GMT+0000
+    uint256 time24 = 1654790400; // Thu Jun 09 2022 16:00:00 GMT+0000
+    uint256 time25 = 1654876800; // Fri Jun 10 2022 16:00:00 GMT+0000
+    uint256 time26 = 1654963200; // Sat Jun 11 2022 16:00:00 GMT+0000
+    uint256 time27 = 1655049600; // Sun Jun 12 2022 16:00:00 GMT+0000
 
     // after cycle
-    uint timeb0 = 1655136000; // Mon Jun 13 2022 16:00:00 GMT+0000
-    uint timeb1 = 1655222400; // Tue Jun 14 2022 16:00:00 GMT+0000
-    uint timeb2 = 1655308800; // Wed Jun 15 2022 16:00:00 GMT+0000
+    uint256 timeb0 = 1655136000; // Mon Jun 13 2022 16:00:00 GMT+0000
+    uint256 timeb1 = 1655222400; // Tue Jun 14 2022 16:00:00 GMT+0000
+    uint256 timeb2 = 1655308800; // Wed Jun 15 2022 16:00:00 GMT+0000
 
     // before cycle
-    uint timec25 = 1657296000; // Fri Jul 08 2022 16:00:00 GMT+0000
-    uint timec26 = 1657382400; // Sat Jul 09 2022 16:00:00 GMT+0000
-    uint timec27 = 1657468800; // Sun Jul 10 2022 16:00:00 GMT+0000
+    uint256 timec25 = 1657296000; // Fri Jul 08 2022 16:00:00 GMT+0000
+    uint256 timec26 = 1657382400; // Sat Jul 09 2022 16:00:00 GMT+0000
+    uint256 timec27 = 1657468800; // Sun Jul 10 2022 16:00:00 GMT+0000
 
     event Transfer(address indexed from, address indexed to, uint256 indexed id);
 
@@ -74,7 +75,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
         moon.tokenURI(0);
     }
 
-    // 
+    //
 
     ////////////////////////////////////////////////
     ////////////////    Mint    ////////////////////
@@ -112,7 +113,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
         moon.mint(address(1), "");
     }
 
-    // 
+    //
 
     ////////////////////////////////////////////////
     ////////////////    Burn    ////////////////////
@@ -165,7 +166,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
         assertEq(moon.getMoonPhaseForDay(3), MoonPhase.WaningGibbous);
         assertEq(moon.getMoonPhaseForDay(4), MoonPhase.WaningGibbous);
         assertEq(moon.getMoonPhaseForDay(5), MoonPhase.WaningGibbous);
-        assertEq(moon.getMoonPhaseForDay(6), MoonPhase.WaningGibbous);        
+        assertEq(moon.getMoonPhaseForDay(6), MoonPhase.WaningGibbous);
         assertEq(moon.getMoonPhaseForDay(7), MoonPhase.LastQuarter);
         assertEq(moon.getMoonPhaseForDay(8), MoonPhase.LastQuarter);
         assertEq(moon.getMoonPhaseForDay(9), MoonPhase.LastQuarter);
@@ -202,7 +203,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
         assertEq(moon.getMoonPhaseForCurrentTime(time3), MoonPhase.WaningGibbous);
         assertEq(moon.getMoonPhaseForCurrentTime(time4), MoonPhase.WaningGibbous);
         assertEq(moon.getMoonPhaseForCurrentTime(time5), MoonPhase.WaningGibbous);
-        assertEq(moon.getMoonPhaseForCurrentTime(time6), MoonPhase.WaningGibbous);        
+        assertEq(moon.getMoonPhaseForCurrentTime(time6), MoonPhase.WaningGibbous);
         assertEq(moon.getMoonPhaseForCurrentTime(time7), MoonPhase.LastQuarter);
         assertEq(moon.getMoonPhaseForCurrentTime(time8), MoonPhase.LastQuarter);
         assertEq(moon.getMoonPhaseForCurrentTime(time9), MoonPhase.LastQuarter);
@@ -235,16 +236,16 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
     }
 
     function testCorrectMoonPhasesForRandomTimes() public {
-        uint rand0 = time0 + (3 * 60 * 60); // 3 hours
+        uint256 rand0 = time0 + (3 * 60 * 60); // 3 hours
         assertEq(moon.getMoonPhaseForCurrentTime(rand0), MoonPhase.FullMoon);
 
-        uint rand1 = time0 + (3 * 60 * 60) + (33 * 60); // 3 hours 33 minutes
+        uint256 rand1 = time0 + (3 * 60 * 60) + (33 * 60); // 3 hours 33 minutes
         assertEq(moon.getMoonPhaseForCurrentTime(rand1), MoonPhase.FullMoon);
 
-        uint rand2 = time0 + (3 * 60 * 60) + (33 * 60) + 13; // 3 hours 33 minutes 13 seconds
+        uint256 rand2 = time0 + (3 * 60 * 60) + (33 * 60) + 13; // 3 hours 33 minutes 13 seconds
         assertEq(moon.getMoonPhaseForCurrentTime(rand2), MoonPhase.FullMoon);
 
-        uint rand3 = time0 + (18 * 60 * 60); // 18 hours
+        uint256 rand3 = time0 + (18 * 60 * 60); // 18 hours
         assertEq(moon.getMoonPhaseForCurrentTime(rand3), MoonPhase.FullMoon);
     }
 
@@ -259,7 +260,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
         uint256 tokenID = moon.mint("secret message");
 
         vm.warp(timeb0);
-        
+
         assertEq(moon.readSecretMessage(tokenID), "secret message");
     }
 
@@ -381,7 +382,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
 
         vm.warp(time14);
 
-        assertEq(moon.castConnexion(tokenID), "secret message");        
+        assertEq(moon.castConnexion(tokenID), "secret message");
     }
 
     function testReadWithCastConnexionWhenCoven() public {
@@ -498,7 +499,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
 
     // TODO sketch mint page in Figma, reach out to designers
     // TODO experiment with how to setup scaffold-eth as the mint page
-    // 
+    //
 
     function onERC721Received(
         address,
@@ -518,7 +519,7 @@ contract MoonInkTest is Test, ERC721TokenReceiver {
         bytes memory b = abi.encodePacked(phaseB);
 
         if (keccak256(a) != keccak256(abi.encodePacked(b))) {
-            emit log            ("Error: a == b not satisfied [bytes]");
+            emit log("Error: a == b not satisfied [bytes]");
             emit log_named_bytes("  Expected", b);
             emit log_named_bytes("    Actual", a);
             fail();
